USB: android: Dont call functionfs_bind during android enable
During composition switch, first sys.usb.config is set
to NONE. As part of this stop adb and android disable
is done. As part of android disable, functionfs_unbind
is not called and it depends on adb release
closed_callback.

On the other side, android enable is getting called.
This calls functionfs bind. At this point functionfs
bind returns with warning because the adb release
closed callback still not finished. Both android
enable and closed callback waits on android mutex
to be acquired. If closed callback gets the mutex
first, warning is not thrown. If android enable
gets the mutext first, warning is thrown. This
is the issue happening in this case.

There seem to be no need to call functionfs bind
in ffs_function_enable as the bind/unbind
should happen in respective ready/closed callbacks.

Change-Id: I4aa75ba7aa2b69c6fa7c701ba05f39263fbf4ad4
Signed-off-by: Sujeet Kumar <ksujeet@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>