net: ipc_router: Reference count server
Everytime a server is searched and then accessed, a global lock is
used to protect the access for its entire duration.

Define a reference element in the server. Get a reference to the server
everytime it is searched and accessed. Put the reference back once the
access is complete. Release the server when the reference count is zero.

Change-Id: I5d2c036f9ae881a11dfc1b1e6ba61f4bee21469e
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>