net: ipc_router: Add support for connect system call
Enable support for connect system call, so that the destination address is
stored as part of the port. Subsequently the clients of IPC Router can use
send and recv system calls in addition to sendto and recvfrom.

Reset any connection between the local port and remote port if the remote
port exits either voluntarily or due to subsystem restart.

Change-Id: Ie931fd312ab7089e91be58857a1f9c1a8f37ca55
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>