usb: tty: Add more buffer space and make the buffer use direct
Due to stall in high throughput in DUN+SoftAP mode, we
needed to improve the way we transfer the buffers.
Here we enlarge the buffer space so we have more space
to buffer the high throughput data and switch to direct
buffer use, excluding the need and the overhead
of a work queue.

Change-Id: I5aa4e70e08574bbe1c6e7e4412530c93751ca5d9
Signed-off-by: Lena Salman <esalman@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>