net: ipc_router: Allocate platform device using platform_device_alloc
Currently platform_device for a service is part of the server structure and
is allocated when the concerned server structure is allocated. This leads
to a problem when the server structure is freed while a reference to the
platform device is held by another thread.

Do not embed the platform device as part of the server structure and
allocate it using platform_device_alloc helper function.

Change-Id: I36284a1d66956d66c8d83234665e4a38ef088ee2
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>