drivers: qcom: qmi: Fix the race condition during QMI handle release
When the QMI client handle is released, the release operation waits only
for the threads whose request has been successfully transmitted and are
waiting for the responses. The release operation does not wait for the
threads whose requests are waiting for the flow control acknowledgement
from the remote side. This causes the QMI client handle to be freed
pre-maturely.

Wait for all the threads in an uninterruptible manner to exit before
freeing the QMI handle.

Change-Id: I4e9b90eb474ed15ed3452ad8ccce46bd9f458ff0
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>