iommu: msm: Avoid null-pointer derefs in ion_map_iommu
Fix the position of the 'out:' label that is jumped to when the size and
alignment checks of the scatter-gather list and iova_length fail. If any
of the three initial checks fail, the goto will invoke a call to:

  ion_iommu_meta_put(iommu_meta)

with iommu_meta as null, resulting in a null-derefence when doing a
kref_put. While here, also handle the case of an alloc failure in
ion_iommu_meta_create() returning an ERR_PTR instead of randomly
crashing when derefrencing it next.

Change-Id: Ib485b8b000ba300357c2f35c0d19f04f9d3460ac
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: http://gerrit.mot.com/693837
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Patrick Auchter <auchter@motorola.com>
Reviewed-by: Fred Fettinger <fettinge@motorola.com>