net: ipc_router: Modify IPC Router flow control algorithm
Currently IPC Router uses stop and wait flow control approach.

Update it to use a variant of sliding-window flow control. Send
the flow control request when the number of packets in flight
reaches the low watermark. Block the transmission when the number
of packets in flight reaches the high watermark.

Change-Id: Ibc422716c646440ea4e965a44dcda81315f66d95
Signed-off-by: Atish Kumar Patra <apatra@codeaurora.org>