net: ipc_router: Fix socket reference in IPC Router
The socket associated with an IPC Router port can be released before
the port is released. This leads to an use after free bug in cases
where the socket is closed while a packet is in flight.

Fix the use after free scenario by holding a sock reference during
creation of the port and releasing that in release port function.

Change-Id: I88794d42b2f9b0d061029fa1170aec8ce772ade4
Signed-off-by: Atish Kumar Patra <apatra@codeaurora.org>