msm: mdss: rotator: ignore panel status for rotator fences
In case of rotator buf sync ioctl, the ioctl could be called on
frame buffer device for which panel is off. However rotator sessions
may still be active. In such cases the call shouldn't fail, thus move
check for panel off only when the target sync fence timeline is for
the panel.

Change-Id: Ic0fa3ccf549a7c4bef206f9a361257f98ad3b4b4
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>