soc: qcom: smd: Catch missed close state transitions
The SMD state machine contains few sync points where it can be ensured that
the remote side has observed a state transition before proceeding to the
next state.  This can lead to situations where the local side misses an
expected state transition and therefore does not perform expected
processing.

If the local side observes a state transition that could have only occurred
if a close state tranition was missed, perform the missed processing.

Change-Id: Ibc4574ffcccc1d3534142327903e41b78dc1e127
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>